pyproject.toml
[project]
name = "garage"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Enzo Lazzarin Dal'osto", email = "wecredsbr@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
